Split types into real and complex versions
Split the PDTs for the interpolators into real and complex versions.
This removes the need for the types to have both real and complex
components at the same time, so that the same name can be used for both
versions of the components. This also reduces the memory overhead of the
types, and makes the fypp templates slightly easier to write. Update the
examples to reflect that the types have been split.

Move the definitions of commonly used fypp variables to
include/common.fypp.

Update the README.