Fix handling of infinities in MTT_init_cp
Due to how infinites are handled by Fortran in complex arithmetic,
intermediate infinities were turned into NaNs, which caused the
MTT_t_test to fail for complex data. Changed so that for the complex
version a check is performed to see if denominator and numerator in the
expression for g_j are zero. If only denom is zero, set a flag to
indicate infinity, and on the next iteration set the fractional part to
0 in the update of g_j. With this fix, all tests are now passing.